Resolve embedding problem with a fully new type
